Update hook pass/fail logic
This complicates the code a bit, but the logic is now the same as other
test frameworks like JUnit.

In order to output diagnostics in a way that makes sense, before_each
and after_each hook output is captured along with the main test function
output.
Capturing hook and test outputs while supporting JUnit-like hook
pass/fail logic requires the output capturing to happen outside of
run_fn.

Now that capturing happens in run_test_functions, all the file
descriptor variable usages happen inside run_test_functions, which makes
it more less side-effecty.

I usually use || for conditional statements so that the statement does
not return unsuccessful if the condition fails, which can be useful if
the script has `set -e`. For example:

    (( fail )) && do_if_failed # if there was no fail, this statement exits 1
    ! (( fail )) || do_if_failed # if there was no fail, this statement exits 0

---

I want the bashaspec-non-tap.sh script to remain simpler than the main
bashaspec.sh script, so I did not copy this change to that script.