Be as portable as possible: support heirloom sh
Least common denominator achieved!

These restrictions make the code another increment uglier, past my
threshold of comfortability. Supporting this is mainly academic because
nobody wants to use something pre-posix anyway. As a result, I've put
this version in alternate-old-versions/bashaspec-ancient.sh

* Heirloom sh does not like the ! keyword for logical negation of
  pipeline status outside of an if condition.
* Heirloom sh does not like $() for command substitution
* Heirloom sh does not like $(()) for arithmetic expansion

The bash source check uses !, but that only happens if the shell is
bash, so it is safe to wrap in an eval.