Remove exit trap; must source after defining tests
When a trap exit handler is being run, `return` with no argument does
not return the exit code of the previous statement anymore, even when
using set -e.

To return the exit code of the previous statement, you need to use
`return $?`.

It was nice to have an exit trap because it allowed sourcing bashaspec
at the top of the test file, but instead of requring users to
`return $?` in all their tests for failure assertions, it is less
cumbersome to require bashaspec to be sourced after defining test
functions.